configure.cmd: print error messages if ./configure not found
    modified:   buildscripts/autotools/configure.cmd
    modified:   buildscripts/configure.cmd

+++ buildscripts/autotools/configure.cmd
@@ -5,5 +5,5 @@ export PATH=$(expr "$PATH" | tr '\\' /)
 
 export ac_executable_extensions=".exe"
 
-test -f ./configure. &&
+test -f ./configure. || { echo "\`./configure' not found !!!"; exit 1; }
 ./configure "$@" 2>&1 | tee configure.log
